requirements_mcp.txt
flask
flask-cors
python-dotenv
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pyautogui
psutil
pillow  # Required for pyautogui screenshots
pathlib-ng
asyncio-extras
typing-extensions 